// src/types/resource.d.ts
export interface Resource {
    id: string;
    name: string;
    content?: string;
    tags?: string[];
    type?: string;
    size?: number;
    fileType?: string;
    uploadDate?: string;
    lastModified?: string;
    [key: string]: any;
    rawData?: string;
}
//# sourceMappingURL=resource.d.ts.map